refactor for different marhsalling
