+++ src/test/scala/onextent/akka/naviblob/azure/ListBlobsSpec.scala
@@ -1,7 +1,7 @@
 package onextent.akka.naviblob.azure
 
 import com.typesafe.scalalogging.LazyLogging
-import onextent.akka.naviblob.azure.avro.AvroStreamReader
+import onextent.akka.naviblob.azure.avro.{AvroStreamReader, EhRecord}
 import onextent.akka.naviblob.azure.storage.{BlobConfig, BlobPaths, Blobber}
 import org.scalatest._
